new/requirements.txt
Django==4.2.6
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-yasg==1.21.5
Pillow==9.5.0
pytest==7.4.3
pytest-django==4.5.2
python-dotenv==1.0.0
requests==2.31.0
# psycopg2==2.9.6
Markdown
